+++ src/main/java/de/samply/converter/ConverterException.java
@@ -1,0 +1,7 @@
+package de.samply.converter;
+
+public class ConverterException extends Exception{
+    public ConverterException(Throwable cause) {
+        super(cause);
+    }
+}