+++ src/main/java/de/samply/fhir/cql/CqlToBundleConverter.java
@@ -1,5 +1,6 @@
 package de.samply.fhir.cql;
 
+import de.samply.converter.ConverterException;
 import de.samply.converter.Format;
 import de.samply.template.ConverterTemplate;
 import org.hl7.fhir.r4.model.Bundle;
@@ -7,8 +8,12 @@ import org.hl7.fhir.r4.model.MeasureReport;
 
 public class CqlToBundleConverter extends CqlRelatedConverter<Bundle> {
 
-    public CqlToBundleConverter(String fhirStoreUrl, String sourceId) {
+    public CqlToBundleConverter(String fhirStoreUrl, String sourceId) throws ConverterException {
         super(fhirStoreUrl, sourceId);
+    }
+
+    public CqlToBundleConverter(String fhirStoreUrl, String sourceId, String httpProxy, String noProxy) throws ConverterException {
+        super(fhirStoreUrl, sourceId, httpProxy, noProxy);
     }
 
     @Override